[Project] Fix #3182 - duplicate file copying with xbuild

+++ main/src/core/MonoDevelop.Core/MonoDevelop.Projects/Project.cs
@@ -393,6 +393,14 @@ namespace MonoDevelop.Projects
 			AddFile (newDir);
 			return newDir;
 		}
+		
+		//HACK: the build code is structured such that support file copying is in here instead of the item handler
+		//so in order to avoid doing them twice when using the msbuild engine, we special-case them
+		bool UsingMSBuildEngine ()
+		{
+			var msbuildHandler = ItemHandler as MonoDevelop.Projects.Formats.MSBuild.MSBuildProjectHandler;
+			return msbuildHandler != null && msbuildHandler.UseXbuild;
+		}
 
 		protected override BuildResult OnBuild (IProgressMonitor monitor, ConfigurationSelector configuration)
 		{
@@ -403,6 +411,15 @@ namespace MonoDevelop.Projects
 				cres.AddError (GettextCatalog.GetString ("Configuration '{0}' not found in project '{1}'", configuration.ToString (), Name));
 				return cres;
 			}
+			
+			StringParserService.Properties["Project"] = Name;
+			
+			if (UsingMSBuildEngine ()) {
+				var r = DoBuild (monitor, configuration);
+				isDirty = false;
+				return r;
+			}
+			
 			string outputDir = conf.OutputDirectory;
 			try {
 				DirectoryInfo directoryInfo = new DirectoryInfo (outputDir);
@@ -415,10 +432,9 @@ namespace MonoDevelop.Projects
 
 			//copy references and files marked to "CopyToOutputDirectory"
 			CopySupportFiles (monitor, configuration);
-
-			StringParserService.Properties["Project"] = Name;
-
+		
 			monitor.BeginTask (GettextCatalog.GetString ("Performing main compilation..."), 0);
+			
 			BuildResult res = DoBuild (monitor, configuration);
 
 			isDirty = false;
@@ -626,15 +642,21 @@ namespace MonoDevelop.Projects
 		protected override void OnClean (IProgressMonitor monitor, ConfigurationSelector configuration)
 		{
 			SetDirty ();
+			
 			ProjectConfiguration config = GetConfiguration (configuration) as ProjectConfiguration;
 			if (config == null) {
 				monitor.ReportError (GettextCatalog.GetString ("Configuration '{0}' not found in project '{1}'", config.Id, Name), null);
 				return;
 			}
 			
-			monitor.BeginTask (GettextCatalog.GetString ("Performing clean..."), 0);
-			// Delete generated files
+			if (UsingMSBuildEngine ()) {
+				DoClean (monitor, config.Selector);
+				return;
+			}
 			
+			monitor.BeginTask (GettextCatalog.GetString ("Performing clean..."), 0);
+			
+			// Delete generated files
 			foreach (FilePath file in GetOutputFiles (configuration)) {
 				if (File.Exists (file)) {
 					FileService.DeleteFile (file);
@@ -642,8 +664,9 @@ namespace MonoDevelop.Projects
 						FileService.DeleteDirectory (file.ParentDirectory);
 				}
 			}
-
+	
 			DeleteSupportFiles (monitor, configuration);
+			
 			DoClean (monitor, config.Selector);
 			monitor.Log.WriteLine ();
 			monitor.Log.WriteLine (GettextCatalog.GetString ("Clean complete"));
